pyproject.toml
[tool.poetry]
name = "quotation_system"
version = "0.1.0"
description = ""
authors = ["Angelo Siqueira <angeloasiqueira@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.9"
python-decouple = "^3.6"
requests = "^2.27.1"
djangorestframework = "^3.13.1"
django-solo = "^2.0.0"
psycopg2-binary = "^2.9.3"
drf-yasg = "^1.20.0"
whitenoise = "^6.0.0"
django-crontab = "^0.7.1"
gunicorn = "^20.1.0"
django-pwa = "^1.0.10"

[tool.poetry.dev-dependencies]
django-extensions = "^3.1.5"
ipython = "^8.2.0"
pytest = "^7.1.2"
pytest-django = "^4.5.2"
model-bakery = "^1.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "quotation_system.settings"
django_debug_mode = true
python_files = "tests.py test_*.py *_tests.py"
addopts = "-x -p no:warnings --strict-markers --no-migrations"
